rename cython_modules
