+++ MANIFEST.in
@@ -1,6 +1,6 @@
-include ms2pip/cython_modules/*.c
-include ms2pip/cython_modules/*.h
-include ms2pip/cython_modules/libms2pip.a
+include ms2pip/ms2pip_c/*.c
+include ms2pip/ms2pip_c/*.h
+include ms2pip/ms2pip_c/libms2pip.a
 include ms2pip/models/*/*.c
 include ms2pip/models/*.h
 include setup.cfg
