+++ setup.cfg
@@ -1,4 +1,4 @@
 [metadata]
 license_file = LICENSE
 [build_clib]
-build_clib = ms2pip/cython_modules
+build_clib = ms2pip/ms2pip_c
